refactor: Move directory_structure handler to file_server module
